refactor so that StorageBase contains the actual token algorithm
